Deprecate ServerSocket class
Also cleanup it and use RubySocket underneath
